requirements.txt
selenium==4.6.1
PyGetWindow~=0.0.9
Pillow~=9.3.0